chore: simplify and fix connection display
