+++ src/dappkit/DappKitSignTxScreen.tsx
@@ -21,7 +21,9 @@ import fontStyles from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
 import variables from 'src/styles/variables'
 import Logger from 'src/utils/Logger'
+import DappsDisclaimer from 'src/walletConnect/screens/DappsDisclaimer'
 import RequestContent from 'src/walletConnect/screens/RequestContent'
+import { useIsDappListed } from 'src/walletConnect/screens/useIsDappListed'
 
 const TAG = 'dappkit/DappKitSignTxScreen'
 
@@ -35,6 +37,7 @@ const DappKitSignTxScreen = ({ route }: Props) => {
   const dappConnectInfo = useSelector(dappConnectInfoSelector)
   const dispatch = useDispatch()
   const { t } = useTranslation()
+  const isDappListed = useIsDappListed(callback)
 
   if (!dappKitRequest) {
     Logger.error(TAG, 'No request found in navigation props')
@@ -93,6 +96,8 @@ const DappKitSignTxScreen = ({ route }: Props) => {
             <CopyIcon />
           </Touchable>
         </View>
+
+        <DappsDisclaimer isDappListed={isDappListed} />
       </RequestContent>
     </SafeAreaView>
   )
